moves functions into generate.js
