+++ tools/cli/commands/generate.js
@@ -3,11 +3,14 @@
  */
 import path from 'path';
 import fs from 'fs';
+import pluralize from 'pluralize';
+import inquirer from 'inquirer';
 
 /**
  * Internal dependencies
  */
-import { promptForGenerate } from '../helpers/promptForProject.js';
+import { promptForType, promptForName } from '../helpers/promptForProject.js';
+import { projectTypes, checkNameValid } from '../helpers/projectHelpers.js';
 import { readPackageJson, writePackageJson } from '../helpers/json';
 import { normalizeGenerateArgv } from '../helpers/normalizeArgv';
 import mergeDirs from '../helpers/mergeDirs';
@@ -33,8 +36,6 @@ export async function generateCli( argv ) {
 	argv = await promptForGenerate( argv );
 	await generateRouter( argv );
 }
-
-/** */
 
 /**
  * Command definition for the generate subcommand.
@@ -68,6 +69,95 @@ export function generateDefine( yargs ) {
 	);
 
 	return yargs;
+}
+
+/**
+ * Prompt for Generating New Project.
+ *
+ * If no project is passed via `options`, then it will prompt for the type of project and the project itself.
+ *
+ * @param {object} options - Passthrough of the argv object.
+ *
+ * @returns {object} argv object with the project property.
+ */
+export async function promptForGenerate( options ) {
+	let typeAnswer = options.type ? { type: options.type } : '';
+	let nameAnswer;
+	let tries = 0;
+
+	// Get project type if not passed as an option.
+	if ( ! typeAnswer || typeAnswer.length === 0 ) {
+		typeAnswer = await promptForType();
+	} else if ( ! projectTypes.includes( pluralize( typeAnswer.type ) ) ) {
+		return new Error( 'Must be a valid project type' );
+	}
+
+	// Get the appropriate list of project prompts based on type
+	const questions = getQuestions( options.type || typeAnswer.type );
+	if ( ! questions ) {
+		return new Error( "Sorry! That's not supported yet!" );
+	}
+
+	// Validate name if it was passed as an option.
+	if ( options.name ) {
+		nameAnswer = checkNameValid( typeAnswer.type, options.name ) ? options.name : null;
+	}
+
+	// Keep asking for name if it fails validation
+	while ( ! nameAnswer ) {
+		try {
+			tries++;
+			const rawNameAnswer = await promptForName();
+			nameAnswer = checkNameValid( typeAnswer.type, rawNameAnswer.name ) ? rawNameAnswer : null;
+		} catch ( err ) {
+			if ( tries >= 3 ) {
+				console.error( 'You are really struggling here. Might be time to take a walk.' );
+				console.error( err.name + ': ' + err.message );
+			}
+		}
+	}
+
+	// Give the list of questions
+	const finalAnswers = await inquirer.prompt( questions );
+
+	return {
+		...options,
+		type: pluralize.singular( typeAnswer.type ),
+		name: nameAnswer.name || options.name,
+		n: nameAnswer.name || options.name,
+		...finalAnswers,
+	};
+}
+
+/**
+ * Returns the appropriate list of questions.
+ *
+ * @param {string} type - The project type. Must be one of projectTypes
+ *
+ * @returns {Array} - Array of questions to ask.
+ */
+export function getQuestions( type ) {
+	const packageQuestions = [
+		{
+			type: 'input',
+			name: 'desc',
+			message: 'Succinctly describe your package:',
+		},
+	];
+	const pluginQuestions = '';
+	const extensionQuestions = '';
+	const githubQuestions = '';
+
+	switch ( pluralize.singular( type ) ) {
+		case 'plugin':
+			return pluginQuestions;
+		case 'package':
+			return packageQuestions;
+		case 'editor-extension':
+			return extensionQuestions;
+		case 'github-action':
+			return githubQuestions;
+	}
 }
 
 /**
